Refactor App component and extract main content; implement fetchListData function and utility functions
