+++ src/api/getListData.ts
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import mockJson from "./mock.json";
+import { getRandom, shuffle, sleep } from "../utils/utils";
 
 export type ListItem = {
   id: number;
@@ -13,38 +14,23 @@ export type DeletedListItem = Omit<ListItem, "description">;
 export const useGetListData = () => {
   const query = useQuery({
     queryKey: ["list"],
-    queryFn: async () => {
-      await sleep(1000);
-
-      if (getRandom() > 85) {
-        console.error("An unexpected error occurred!");
-        throw new Error("👀");
-      }
-
-      const mockData = mockJson as Omit<ListItem, "isVisible">[];
-
-      return shuffle(mockData).map((item) => {
-        return { ...item, isVisible: getRandom() > 50 ? true : false };
-      });
-    },
+    queryFn: fetchListData,
   });
 
   return query;
 };
 
-const getRandom = () => Math.floor(Math.random() * 100);
+const fetchListData = async () => {
+  await sleep(1000);
 
-const sleep = (ms: number) => {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const shuffle = <T extends any[]>(array: T): T => {
-  for (let i = array.length - 1; i >= 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    const temp = array[i];
-    array[i] = array[j];
-    array[j] = temp;
+  if (getRandom() > 85) {
+    console.error("An unexpected error occurred!");
+    throw new Error("👀");
   }
-  return array;
+
+  const mockData = mockJson as Omit<ListItem, "isVisible">[];
+
+  return shuffle(mockData).map((item) => {
+    return { ...item, isVisible: getRandom() > 50 ? true : false };
+  });
 };
