+++ src/utils/utils.ts
@@ -1,0 +1,16 @@
+export const getRandom = () => Math.floor(Math.random() * 100);
+
+export const sleep = (ms: number) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const shuffle = <T extends any[]>(array: T): T => {
+  for (let i = array.length - 1; i >= 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = array[i];
+    array[i] = array[j];
+    array[j] = temp;
+  }
+  return array;
+};
